requirements-docker.txt
# Minimal requirements for Airflow Docker container
# Core Airflow is already installed in the base image

# Reddit API
praw==7.7.1

# AWS Services  
boto3==1.34.0
pyarrow==14.0.2

# Data Processing
pandas==2.1.4

# Configuration
python-dotenv==1.0.0

# Airflow Providers (compatible versions)
apache-airflow-providers-amazon>=7.0.0
apache-airflow-providers-postgres>=5.0.0